+++ test.txt
@@ -1,0 +1,2 @@
+I am van, I am an artist, I am an performance artist. 
+I am going to fulfill their fantasy, the deep dark fantasy
